Initial example for C++ Contracts.
This initial test creates two simple libraries, from the same source, that use enforce of ignore semantics. And creates a matrix of the semantic of enforce/ignore for the main source vs library, and for dynamic or static linking of the library.

The result for this is that *only* the enforce compiled library will terminate the program.